requirements.txt
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
python-dotenv==1.0.1
tiktoken==0.7.0
numpy==1.26.4
langchain-chroma